Finally move SGP surrogate traits in MOE
