+++ gp/src/lib.rs
@@ -20,11 +20,8 @@ mod parameters;
 mod sgp_parameters;
 mod utils;
 
-mod gp_surrogate;
-
 pub use algorithm::*;
 pub use errors::*;
-pub use gp_surrogate::*;
 pub use parameters::*;
 pub use sgp_algorithm::*;
 pub use sgp_parameters::*;
